backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.1
PyMySQL==1.1.1
SQLAlchemy==2.0.32
alembic==1.13.2
redis==5.0.7
PyJWT==2.9.0

